requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-braces==1.15.0
django-environ==0.8.1
django-oauth-toolkit==2.0.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jwcrypto==1.3.1
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
